Add Timeline 1.8.0 Samples
**What Changed and Why?**
Added Timeline Samples showing off how to make custom tracks, as well as a fully animated cutscene to show you how it's done!

The most important lesson here is that creating cutscenes is hard, so we should plan it out from the beginning. As well, resources have been provided in the github wiki to Unity courses that teach you how to create a cutscene from start to finish.

Signed-off-by: CaseyDeCoder <hofland.casey@gmail.com>