Update NamingValidator
**What Changed and Why?**
The naming validator was bumped to the latest version, now adding support for validating folder names.

Additionally, the A_ prefix allowance for models has been removed. Instead, a model that is purely brought in for its animation is expected to be of type SK_ or ST_ based on if it is rigged or not.

Signed-off-by: CaseyDeCoder <hofland.casey@gmail.com>