Improve FMOD QoL inside Unity
**What Changed and Why?**
Added Presets for the StudioBankLoader and StudioEventEmitter so that, by default, these will play and stop during OnEnable and OnDisable.

Additionally, the Unity Player has been instructed to run in the background. This allows for audio to keep playing when you defocus the Unity Editor, severely improving the experience for FMOD Live Update!

Signed-off-by: CaseyDeCoder <hofland.casey@gmail.com>